WIP 5

+++ install.sh
@@ -12,9 +12,12 @@
 # 5. Add on boot script for automatically running miner
 # 6. Add script for ease of starting miner
 
-installdir='miner'
-currentuser='pi'
-logfile='/home/pi/gspmi-install.log'
+installdir="miner"
+currentuser="pi"
+
+home=$(echo ~)
+
+logfile="$home/gspmi-install.log"
 
 info='\n \e[46m - \e[49m'
 ok='\n \e[42m + \e[49m'
@@ -30,19 +33,19 @@ fi
 echo -e "$info This script will download and compile the miner and configure the miner to start on boot."
 echo -e "$info This may take a long time to run ~10-15 minutes. Do not quit or turn off your device while this is installing!\n"
 
-# read -p "Are you sure you want to install? (y/n) " -n 1 -r
-# echo -e
-# if [[ ! $REPLY =~ ^[Yy]$ ]]; then
-#   exit 1
-# fi
+read -p "Are you sure you want to install? (y/n) " -n 1 -r
+echo -e
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+  exit 1
+fi
 
 # Install requirements
-echo -e "$info updating and installing packages"
+echo -e "$info Updating and installing packages"
 
 sudo apt-get update -y >> $logfile 2>&1
 sudo apt-get install git screen libcurl4-openssl-dev pkg-config libtool libudev-dev libncurses5-dev -y >> $logfile 2>&1
 
-echo -e "$ok packages updated and installed"
+echo -e "$ok Packages updated and installed"
 
 # Remove any old attempts
 rm -rf ~/$installdir >> $logfile 2>&1
@@ -51,35 +54,35 @@ rm -rf ~/$installdir >> $logfile 2>&1
 mkdir ~/$installdir && cd ~/$installdir
 
 # Grab file
-echo -e "$info downloading configure files and scripts"
+echo -e "$info Downloading configure files and scripts"
 
 curl -o config.json -l https://raw.githubusercontent.com/davidmaitland/Gridseed-PiMiner/master/config.json >> $logfile 2>&1
 
-echo -e "$ok files downloaded"
+echo -e "$ok Files downloaded"
 
 # Clone repo
-echo -e "$info cloning dtbartle's version of cgminer"
+echo -e "$info Cloning dtbartle's version of cgminer"
 
 git clone https://github.com/dtbartle/cgminer-gc3355.git >> $logfile 2>&1
 
-echo -e "$ok cloned"
+echo -e "$ok Cloned repo"
 
 cd ./cgminer-gc3355
 
-echo -e "$info configuring cgminer. This may take a while. Time for a beer?"
+echo -e "$info Configuring cgminer. This may take a while. Time for a beer?"
 
 ./configure --enable-scrypt --enable-gridseed >> $logfile 2>&1
 
-echo -e "$ok configured cgminer"
+echo -e "$ok Configured cgminer"
 
-echo -e "$info compiling cgminer. This will take even longer."
+echo -e "$info Compiling cgminer. This will take even longer."
 
 make >> $logfile 2>&1
 
-echo -e "$ok complied cgminer! Put that beer down!"
+echo -e "$ok Complied cgminer! Put that beer down!"
 
-echo -e "$info installing cgminer"
+echo -e "$info Installing cgminer"
 
 sudo make install >> $logfile 2>&1
 
-echo -e "$ok installed cgminer"
+echo -e "$ok Installed cgminer"
